+++ R/forplum.R
@@ -109,6 +109,7 @@ draw.pbmeasured <- function(set=get('info'), rotate.axes=FALSE, rev.d=FALSE, rev
 #' @return A plot of the modelled (and optionally the measured) 210Pb values
 #' @export
 draw.pbmodelled <- function(set=get('info'), BCAD=set$BCAD, rotate.axes=FALSE, rev.d=FALSE, rev.age=FALSE, pb.lim=c(), d.lim=c(), d.lab=c(), pb.lab=c(), pbmodelled.col=function(x) rgb(0,0,1,x), pbmeasured.col="blue", supp.col="purple", plot.measured=TRUE, age.lim=c(), mgp=mgp, pb.lty=1) {
+
   pb <- set$dets[set$dets[,9] == 5,]
   depths <- pb[,4] # set$detsOrig[,2]
   dns <- pb[,6] # set$detsOrig[,3]
@@ -170,10 +171,8 @@ draw.pbmodelled <- function(set=get('info'), BCAD=set$BCAD, rotate.axes=FALSE, r
     # save the values for later
     set$Ai <- Ai
     set$A.rng <- A.rng
-    #assign_to_global("info", set, .GlobalEnv) # doesn't work
 
     this <- ifelse(rotate.axes, 3, 4)
-    pretty.pb <- pretty(c(pbmin, pbmax)) # not OK?
     pretty.pb <- pretty(pb.lim)
     onbp <- pb2bp(pretty.pb)
     if(BCAD)
@@ -183,32 +182,30 @@ draw.pbmodelled <- function(set=get('info'), BCAD=set$BCAD, rotate.axes=FALSE, r
     mtext(pb.lab, this, 2.5, col=pbmeasured.col, cex=.8)
 
     for(i in 1:length(depths)) {
-      if(BCAD) {
-        ages <- pb2bp(rev(Ai$x[[i]]))
-        z <- t(rev(Ai$y[[i]]))/hght
-      } else {
-          ages <- pb2bp(Ai$x[[i]])
-          z <- t(Ai$y[[i]])/hght
-        }
-
-      if(rotate.axes)
-        ghost.mirror(ages, c(depths[i]-thickness[i], depths[i]), t(z), col=pbmodelled.col(seq(0, 1-max(z),  length=50))) else
-          ghost.mirror(c(depths[i]-thickness[i], depths[i]), ages, z, col=pbmodelled.col(seq(0, 1-max(z), length=50)))
+      ages <- pb2bp(Ai$x[[i]], AD=BCAD)
+      if(BCAD) 
+		ages <- rev(ages)
+      z <- matrix(Ai$y[[i]]/hght, nrow=1)
+      d_slice <- c(depths[i]-thickness[i], depths[i])
+      modelled_col <- pbmodelled.col(seq(0, 1-max(z), length=50))
+	  
+	  # we're not using ghost.mirror, because of BC/AD axis reversal issues
+      if(rotate.axes) {
+        if(BCAD)	  
+          image(ages, d_slice, z, add=TRUE, col=modelled_col, useRaster=FALSE) else
+            image(ages, d_slice, z, add=TRUE, col=modelled_col, useRaster=FALSE)
+        } else {
+	     if(BCAD)  
+  	       image(d_slice, ages, z, add=TRUE, col=modelled_col, useRaster=FALSE) else
+	         image(d_slice, ages, z, add=TRUE, col=modelled_col, useRaster=FALSE)
+        }	  
     }
   }
 
-#   # indicate in redscale which Pb-210 data have most likely reached background
-#   if(draw.background) {
-#     bg <- background(set)
-#     set$background <- bg
-#     assign_to_global("info", set, .GlobalEnv)
-#     if(rotate.axes)
-#       abline(h=set$dets[,4]-(set$dets[,5]/2), col=rgb(bg,0,0,bg), lty=3, lwd=bg) else
-#         abline(v=set$dets[,4]-(set$dets[,5]/2), col=rgb(bg,0,0,bg), lty=3, lwd=bg)
-#   }
-
-    if(plot.measured)
-      draw.pbmeasured(set=set, newplot=FALSE, rotate.axes=rotate.axes, BCAD=BCAD, on.agescale=TRUE, pb.lim=pb.lim, age.lim=age.lim, supp.col=supp.col)	
+  if(plot.measured)
+    draw.pbmeasured(set=set, newplot=FALSE, rotate.axes=rotate.axes, BCAD=BCAD, on.agescale=TRUE, pb.lim=pb.lim, age.lim=age.lim, supp.col=supp.col)	
+  
+  invisible(set)
 }